Refactor Cloud service callback handling (#7)
Refactor the callback handling to avoid the 3-level callback architecture. The background publish callback is a lamdba to bind the publish details and original payload, which will add a handler for the expected ack that calls the user callback, or prepares the user callback to be invoked in the next main loop invocation, if full acknowledgement is not required. The original payload data is copied only once, and moved using the move constructor to String.

Command callback handlers are separated from ack handlers, so parameters invalid for one or the other don't need to be assigned unused values.